Sync updates from stainless branch: hardikjshah/dev
