+++ src/llama_stack_client/resources/embeddings.py
@@ -18,7 +18,7 @@ from .._response import (
     async_to_streamed_response_wrapper,
 )
 from .._base_client import make_request_options
-from ..types.embeddings_response import EmbeddingsResponse
+from ..types.create_embeddings_response import CreateEmbeddingsResponse
 
 __all__ = ["EmbeddingsResource", "AsyncEmbeddingsResource"]
 
@@ -57,7 +57,7 @@ class EmbeddingsResource(SyncAPIResource):
         extra_query: Query | None = None,
         extra_body: Body | None = None,
         timeout: float | httpx.Timeout | None | NotGiven = NOT_GIVEN,
-    ) -> EmbeddingsResponse:
+    ) -> CreateEmbeddingsResponse:
         """
         Generate OpenAI-compatible embeddings for the given input using the specified
         model.
@@ -101,7 +101,7 @@ class EmbeddingsResource(SyncAPIResource):
             options=make_request_options(
                 extra_headers=extra_headers, extra_query=extra_query, extra_body=extra_body, timeout=timeout
             ),
-            cast_to=EmbeddingsResponse,
+            cast_to=CreateEmbeddingsResponse,
         )
 
 
@@ -139,7 +139,7 @@ class AsyncEmbeddingsResource(AsyncAPIResource):
         extra_query: Query | None = None,
         extra_body: Body | None = None,
         timeout: float | httpx.Timeout | None | NotGiven = NOT_GIVEN,
-    ) -> EmbeddingsResponse:
+    ) -> CreateEmbeddingsResponse:
         """
         Generate OpenAI-compatible embeddings for the given input using the specified
         model.
@@ -183,7 +183,7 @@ class AsyncEmbeddingsResource(AsyncAPIResource):
             options=make_request_options(
                 extra_headers=extra_headers, extra_query=extra_query, extra_body=extra_body, timeout=timeout
             ),
-            cast_to=EmbeddingsResponse,
+            cast_to=CreateEmbeddingsResponse,
         )
